LVGL setup with basic display states. Created component for misc src files including pir and buzzer. Began development of kiosk state machine
