+++ src/app/display/display_frames.c
@@ -1,356 +1,69 @@
-#include "esp_lcd_gc9a01.h"
-#include <stdio.h>
-#include <unistd.h>
-#include <sys/lock.h>
-#include <sys/param.h>
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-#include "esp_timer.h"
-#include "esp_lcd_panel_io.h"
-#include "esp_lcd_panel_vendor.h"
-#include "esp_lcd_panel_ops.h"
-#include "driver/gpio.h"
-#include "driver/spi_master.h"
-#include "esp_err.h"
-#include "esp_log.h"
-#include "lvgl.h"
 #include "display_frames.h"
 
-#include <inttypes.h>
-
-#include "freertos/semphr.h"
-#include "esp_heap_caps.h"
-#include "esp_lcd_panel_io_interface.h"
-
-#define EXAMPLE_LCD_HOST SPI2_HOST
-#define EXAMPLE_LCD_PIXEL_CLOCK_HZ (20 * 1000 * 1000)
-#define TEST_LCD_BIT_PER_PIXEL 16
-#define EXAMPLE_LCD_BK_LIGHT_ON_LEVEL 1
-#define EXAMPLE_LCD_BK_LIGHT_OFF_LEVEL !EXAMPLE_LCD_BK_LIGHT_ON_LEVEL
-#define EXAMPLE_PIN_NUM_SCLK 10
-#define EXAMPLE_PIN_NUM_MOSI 7
-#define EXAMPLE_PIN_NUM_MISO 2
-#define EXAMPLE_PIN_NUM_LCD_DC 18
-#define EXAMPLE_PIN_NUM_LCD_RST 11 // Will need to change this to a different pin
-#define EXAMPLE_PIN_NUM_LCD_CS 15
-#define EXAMPLE_PIN_NUM_BK_LIGHT 2
-#define EXAMPLE_PIN_NUM_TOUCH_CS 15
-
-#define EXAMPLE_LCD_H_RES 240
-#define EXAMPLE_LCD_V_RES 240
-// Bit number used to represent command and parameter
-#define EXAMPLE_LCD_CMD_BITS 8
-#define EXAMPLE_LCD_PARAM_BITS 8
-#define TEST_DELAY_TIME_MS (3000)
-
-#define EXAMPLE_LVGL_DRAW_BUF_LINES 20 // number of display lines in each draw buffer
-#define EXAMPLE_LVGL_TICK_PERIOD_MS 2
-#define EXAMPLE_LVGL_TASK_MAX_DELAY_MS 500
-#define EXAMPLE_LVGL_TASK_MIN_DELAY_MS 1
-#define EXAMPLE_LVGL_TASK_STACK_SIZE (4 * 1024)
-#define EXAMPLE_LVGL_TASK_PRIORITY 2
-
-#define CONFIG_EXAMPLE_LCD_CONTROLLER_GC9A01 1
-static const char *DISPLAY_TAG = "example";
-
-// // LVGL library is not thread-safe, this example will call LVGL APIs from different tasks, so use a mutex to protect it
-static _lock_t lvgl_api_lock;
-
-extern void example_lvgl_demo_ui(lv_disp_t *disp);
-
-static bool example_notify_lvgl_flush_ready(esp_lcd_panel_io_handle_t panel_io, esp_lcd_panel_io_event_data_t *edata, void *user_ctx)
+lv_obj_t *display_idle(lv_display_t *disp)
 {
-    lv_display_t *disp = (lv_display_t *)user_ctx;
-    lv_display_flush_ready(disp);
-    return false;
+    lv_obj_t *scr = lv_display_get_screen_active(disp);
+
+    static lv_style_t style;
+    lv_style_init(&style);
+    lv_style_set_text_font(&style, &lv_font_montserrat_32);
+
+    lv_obj_t *label1 = lv_label_create(scr);
+    lv_obj_add_style(label1, &style, 0);
+    lv_obj_set_style_bg_color(scr, lv_palette_main(LV_PALETTE_BLUE), 0);
+    lv_label_set_text(label1, "Scan a valid\nID Card");
+
+    lv_obj_align(label1, LV_ALIGN_CENTER, 0, 0);
+
+    return label1;
+}
+lv_obj_t *display_check_in_failed(lv_display_t *disp)
+{
+    lv_obj_t *scr = lv_display_get_screen_active(disp);
+    static lv_style_t style;
+
+    lv_style_init(&style);
+    lv_style_set_text_font(&style, &lv_font_montserrat_32);
+
+    lv_obj_t *label1 = lv_label_create(scr);
+    lv_obj_add_style(label1, &style, 0);
+    lv_obj_set_style_bg_color(scr, lv_palette_main(LV_PALETTE_RED), 0);
+    lv_label_set_text(label1, "Check-In\nFailed");
+
+    lv_obj_align(label1, LV_ALIGN_CENTER, 0, 0);
+
+    return label1;
+}
+lv_obj_t *display_check_in_success(lv_display_t *disp)
+{
+    lv_obj_t *scr = lv_display_get_screen_active(disp);
+    static lv_style_t style;
+    lv_style_init(&style);
+    lv_style_set_text_font(&style, &lv_font_montserrat_32);
+
+    lv_obj_t *label1 = lv_label_create(scr);
+    lv_obj_add_style(label1, &style, 0);
+    lv_obj_set_style_bg_color(scr, lv_palette_main(LV_PALETTE_GREEN), 0);
+    lv_label_set_text(label1, "Check-In\nSuccess");
+
+    lv_obj_align(label1, LV_ALIGN_CENTER, 0, 0);
+
+    return label1;
 }
 
-/* Rotate display and touch, when rotated screen in LVGL. Called when driver parameters are updated. */
-static void example_lvgl_port_update_callback(lv_display_t *disp)
+lv_obj_t *display_transmitting(lv_display_t *disp)
 {
-    esp_lcd_panel_handle_t panel_handle = lv_display_get_user_data(disp);
-    lv_display_rotation_t rotation = lv_display_get_rotation(disp);
+    lv_obj_t *scr = lv_display_get_screen_active(disp);
+    static lv_style_t style;
+    lv_style_init(&style);
+    lv_style_set_text_font(&style, &lv_font_montserrat_32);
 
-    switch (rotation)
-    {
-    case LV_DISPLAY_ROTATION_0:
-        // Rotate LCD display
-        esp_lcd_panel_swap_xy(panel_handle, false);
-        esp_lcd_panel_mirror(panel_handle, true, false);
-        break;
-    case LV_DISPLAY_ROTATION_90:
-        // Rotate LCD display
-        esp_lcd_panel_swap_xy(panel_handle, true);
-        esp_lcd_panel_mirror(panel_handle, true, true);
-        break;
-    case LV_DISPLAY_ROTATION_180:
-        // Rotate LCD display
-        esp_lcd_panel_swap_xy(panel_handle, false);
-        esp_lcd_panel_mirror(panel_handle, false, true);
-        break;
-    case LV_DISPLAY_ROTATION_270:
-        // Rotate LCD display
-        esp_lcd_panel_swap_xy(panel_handle, true);
-        esp_lcd_panel_mirror(panel_handle, false, false);
-        break;
-    }
-}
+    lv_obj_t *label1 = lv_label_create(scr);
+    lv_obj_add_style(label1, &style, 0);
+    lv_obj_set_style_bg_color(scr, lv_palette_main(LV_PALETTE_YELLOW), 0);
+    lv_label_set_text(label1, "Checking\nDatabase");
 
-static void example_lvgl_flush_cb(lv_display_t *disp, const lv_area_t *area, uint8_t *px_map)
-{
-    example_lvgl_port_update_callback(disp);
-    esp_lcd_panel_handle_t panel_handle = lv_display_get_user_data(disp);
-    int offsetx1 = area->x1;
-    int offsetx2 = area->x2;
-    int offsety1 = area->y1;
-    int offsety2 = area->y2;
-    // because SPI LCD is big-endian, we need to swap the RGB bytes order
-    lv_draw_sw_rgb565_swap(px_map, (offsetx2 + 1 - offsetx1) * (offsety2 + 1 - offsety1));
-    // copy a buffer's content to a specific area of the display
-    esp_lcd_panel_draw_bitmap(panel_handle, offsetx1, offsety1, offsetx2 + 1, offsety2 + 1, px_map);
-}
+    lv_obj_align(label1, LV_ALIGN_CENTER, 0, 0);
 
-static void example_increase_lvgl_tick(void *arg)
-{
-    /* Tell LVGL how many milliseconds has elapsed */
-    lv_tick_inc(EXAMPLE_LVGL_TICK_PERIOD_MS);
-}
-
-static void example_lvgl_port_task(void *arg)
-{
-    ESP_LOGI(DISPLAY_TAG, "Starting LVGL task");
-    uint32_t time_till_next_ms = 0;
-    uint32_t time_threshold_ms = 1000 / CONFIG_FREERTOS_HZ;
-    while (1)
-    {
-        _lock_acquire(&lvgl_api_lock);
-        time_till_next_ms = lv_timer_handler();
-        _lock_release(&lvgl_api_lock);
-        // in case of triggering a task watch dog time out
-        time_till_next_ms = MAX(time_till_next_ms, time_threshold_ms);
-        usleep(1000 * time_till_next_ms);
-    }
-}
-
-static SemaphoreHandle_t refresh_finish = NULL;
-
-IRAM_ATTR static bool test_notify_refresh_ready(esp_lcd_panel_io_handle_t panel_io, esp_lcd_panel_io_event_data_t *edata, void *user_ctx)
-{
-    BaseType_t need_yield = pdFALSE;
-
-    xSemaphoreGiveFromISR(refresh_finish, &need_yield);
-    return (need_yield == pdTRUE);
-}
-
-static void test_draw_bitmap(esp_lcd_panel_handle_t panel_handle)
-{
-    refresh_finish = xSemaphoreCreateBinary();
-
-    uint16_t row_line = EXAMPLE_LCD_V_RES / TEST_LCD_BIT_PER_PIXEL;
-    uint8_t byte_per_pixel = TEST_LCD_BIT_PER_PIXEL / 8;
-    uint8_t *color = (uint8_t *)heap_caps_calloc(1, row_line * EXAMPLE_LCD_H_RES * byte_per_pixel, MALLOC_CAP_DMA);
-
-    for (int j = 0; j < TEST_LCD_BIT_PER_PIXEL; j++)
-    {
-        for (int i = 0; i < row_line * EXAMPLE_LCD_H_RES; i++)
-        {
-            for (int k = 0; k < byte_per_pixel; k++)
-            {
-                color[i * byte_per_pixel + k] = (SPI_SWAP_DATA_TX(BIT(j), TEST_LCD_BIT_PER_PIXEL) >> (k * 8)) & 0xff;
-            }
-        }
-        esp_lcd_panel_draw_bitmap(panel_handle, 0, j * row_line, EXAMPLE_LCD_H_RES, (j + 1) * row_line, color);
-        xSemaphoreTake(refresh_finish, portMAX_DELAY);
-    }
-    free(color);
-    vSemaphoreDelete(refresh_finish);
-}
-
-void display_test_1(void *pvParameters)
-{
-    ESP_LOGI(DISPLAY_TAG, "Initialize SPI bus");
-    const spi_bus_config_t buscfg = GC9A01_PANEL_BUS_SPI_CONFIG(EXAMPLE_PIN_NUM_SCLK, EXAMPLE_PIN_NUM_MOSI,
-                                                                EXAMPLE_LCD_H_RES * 80 * TEST_LCD_BIT_PER_PIXEL / 8);
-    spi_bus_initialize(EXAMPLE_LCD_HOST, &buscfg, SPI_DMA_CH_AUTO);
-
-    ESP_LOGI(DISPLAY_TAG, "Install panel IO");
-    esp_lcd_panel_io_handle_t io_handle = NULL;
-    const esp_lcd_panel_io_spi_config_t io_config = GC9A01_PANEL_IO_SPI_CONFIG(EXAMPLE_PIN_NUM_LCD_CS, EXAMPLE_PIN_NUM_LCD_DC,
-                                                                               test_notify_refresh_ready, NULL);
-    // Attach the LCD to the SPI bus
-    esp_lcd_new_panel_io_spi((esp_lcd_spi_bus_handle_t)EXAMPLE_LCD_HOST, &io_config, &io_handle);
-
-    ESP_LOGI(DISPLAY_TAG, "Install gc9a01 panel driver");
-    esp_lcd_panel_handle_t panel_handle = NULL;
-    const esp_lcd_panel_dev_config_t panel_config = {
-        .reset_gpio_num = EXAMPLE_PIN_NUM_LCD_RST,
-#if ESP_IDF_VERSION < ESP_IDF_VERSION_VAL(5, 0, 0)
-        .color_space = ESP_LCD_COLOR_SPACE_BGR,
-#else
-        .rgb_endian = LCD_RGB_ENDIAN_BGR,
-#endif
-        .bits_per_pixel = TEST_LCD_BIT_PER_PIXEL,
-    };
-    esp_lcd_new_panel_gc9a01(io_handle, &panel_config, &panel_handle);
-    esp_lcd_panel_reset(panel_handle);
-    esp_lcd_panel_init(panel_handle);
-    esp_lcd_panel_invert_color(panel_handle, true);
-    esp_lcd_panel_mirror(panel_handle, true, false);
-#if ESP_IDF_VERSION < ESP_IDF_VERSION_VAL(5, 0, 0)
-    esp_lcd_panel_disp_off(panel_handle, false);
-#else
-    esp_lcd_panel_disp_on_off(panel_handle, true);
-#endif
-
-    test_draw_bitmap(panel_handle);
-    vTaskDelay(pdMS_TO_TICKS(TEST_DELAY_TIME_MS));
-
-    esp_lcd_panel_del(panel_handle);
-    esp_lcd_panel_io_del(io_handle);
-    spi_bus_free(EXAMPLE_LCD_HOST);
-
-    vTaskDelete(NULL);
-}
-
-void display_test(void *pvParameters)
-{
-    // ESP_LOGI(DISPLAY_TAG, "Turn off LCD backlight");
-    // gpio_config_t bk_gpio_config = {
-    //     .mode = GPIO_MODE_OUTPUT,
-    //     .pin_bit_mask = 1ULL << EXAMPLE_PIN_NUM_BK_LIGHT};
-    // ESP_ERROR_CHECK(gpio_config(&bk_gpio_config));
-
-    ESP_LOGI(DISPLAY_TAG, "Initialize SPI bus");
-    spi_bus_config_t buscfg = {
-        .sclk_io_num = EXAMPLE_PIN_NUM_SCLK,
-        .mosi_io_num = EXAMPLE_PIN_NUM_MOSI,
-        .miso_io_num = -1,
-        .quadwp_io_num = -1,
-        .quadhd_io_num = -1,
-        .max_transfer_sz = EXAMPLE_LCD_H_RES * 80 * sizeof(uint16_t),
-    };
-    ESP_ERROR_CHECK(spi_bus_initialize(EXAMPLE_LCD_HOST, &buscfg, SPI_DMA_CH_AUTO));
-
-    ESP_LOGI(DISPLAY_TAG, "Install panel IO");
-    esp_lcd_panel_io_handle_t io_handle = NULL;
-    esp_lcd_panel_io_spi_config_t io_config = {
-        .dc_gpio_num = EXAMPLE_PIN_NUM_LCD_DC,
-        .cs_gpio_num = EXAMPLE_PIN_NUM_LCD_CS,
-        .pclk_hz = EXAMPLE_LCD_PIXEL_CLOCK_HZ,
-        .lcd_cmd_bits = EXAMPLE_LCD_CMD_BITS,
-        .lcd_param_bits = EXAMPLE_LCD_PARAM_BITS,
-        .spi_mode = 0,
-        .trans_queue_depth = 10,
-    };
-    // Attach the LCD to the SPI bus
-    ESP_ERROR_CHECK(esp_lcd_new_panel_io_spi((esp_lcd_spi_bus_handle_t)EXAMPLE_LCD_HOST, &io_config, &io_handle));
-
-    esp_lcd_panel_handle_t panel_handle = NULL;
-    esp_lcd_panel_dev_config_t panel_config = {
-        .reset_gpio_num = EXAMPLE_PIN_NUM_LCD_RST,
-        .rgb_ele_order = LCD_RGB_ELEMENT_ORDER_BGR,
-        .bits_per_pixel = 16,
-    };
-#if CONFIG_EXAMPLE_LCD_CONTROLLER_ILI9341
-    ESP_LOGI(DISPLAY_TAG, "Install ILI9341 panel driver");
-    ESP_ERROR_CHECK(esp_lcd_new_panel_ili9341(io_handle, &panel_config, &panel_handle));
-#elif CONFIG_EXAMPLE_LCD_CONTROLLER_GC9A01
-    ESP_LOGI(DISPLAY_TAG, "Install GC9A01 panel driver");
-    ESP_ERROR_CHECK(esp_lcd_new_panel_gc9a01(io_handle, &panel_config, &panel_handle));
-#endif
-
-    ESP_ERROR_CHECK(esp_lcd_panel_reset(panel_handle));
-    ESP_ERROR_CHECK(esp_lcd_panel_init(panel_handle));
-#if CONFIG_EXAMPLE_LCD_CONTROLLER_GC9A01
-    ESP_ERROR_CHECK(esp_lcd_panel_invert_color(panel_handle, true));
-#endif
-    ESP_ERROR_CHECK(esp_lcd_panel_mirror(panel_handle, true, false));
-
-    // user can flush pre-defined pattern to the screen before we turn on the screen or backlight
-    ESP_ERROR_CHECK(esp_lcd_panel_disp_on_off(panel_handle, true));
-
-    // ESP_LOGI(DISPLAY_TAG, "Turn on LCD backlight");
-    // gpio_set_level(EXAMPLE_PIN_NUM_BK_LIGHT, EXAMPLE_LCD_BK_LIGHT_ON_LEVEL);
-
-    ESP_LOGI(DISPLAY_TAG, "Initialize LVGL library");
-    lv_init();
-
-    // create a lvgl display
-    lv_display_t *display = lv_display_create(EXAMPLE_LCD_H_RES, EXAMPLE_LCD_V_RES);
-
-    // alloc draw buffers used by LVGL
-    // it's recommended to choose the size of the draw buffer(s) to be at least 1/10 screen sized
-    size_t draw_buffer_sz = EXAMPLE_LCD_H_RES * EXAMPLE_LVGL_DRAW_BUF_LINES * sizeof(lv_color16_t);
-
-    void *buf1 = spi_bus_dma_memory_alloc(EXAMPLE_LCD_HOST, draw_buffer_sz, 0);
-    assert(buf1);
-    void *buf2 = spi_bus_dma_memory_alloc(EXAMPLE_LCD_HOST, draw_buffer_sz, 0);
-    assert(buf2);
-    // initialize LVGL draw buffers
-    lv_display_set_buffers(display, buf1, buf2, draw_buffer_sz, LV_DISPLAY_RENDER_MODE_PARTIAL);
-    // associate the mipi panel handle to the display
-    lv_display_set_user_data(display, panel_handle);
-    // set color depth
-    lv_display_set_color_format(display, LV_COLOR_FORMAT_RGB565);
-    // set the callback which can copy the rendered image to an area of the display
-    lv_display_set_flush_cb(display, example_lvgl_flush_cb);
-
-    ESP_LOGI(DISPLAY_TAG, "Install LVGL tick timer");
-    // Tick interface for LVGL (using esp_timer to generate 2ms periodic event)
-    const esp_timer_create_args_t lvgl_tick_timer_args = {
-        .callback = &example_increase_lvgl_tick,
-        .name = "lvgl_tick"};
-    esp_timer_handle_t lvgl_tick_timer = NULL;
-    ESP_ERROR_CHECK(esp_timer_create(&lvgl_tick_timer_args, &lvgl_tick_timer));
-    ESP_ERROR_CHECK(esp_timer_start_periodic(lvgl_tick_timer, EXAMPLE_LVGL_TICK_PERIOD_MS * 1000));
-
-    ESP_LOGI(DISPLAY_TAG, "Register io panel event callback for LVGL flush ready notification");
-    const esp_lcd_panel_io_callbacks_t cbs = {
-        .on_color_trans_done = example_notify_lvgl_flush_ready,
-    };
-    /* Register done callback */
-    ESP_ERROR_CHECK(esp_lcd_panel_io_register_event_callbacks(io_handle, &cbs, display));
-
-#if CONFIG_EXAMPLE_LCD_TOUCH_ENABLED
-    esp_lcd_panel_io_handle_t tp_io_handle = NULL;
-    esp_lcd_panel_io_spi_config_t tp_io_config = ESP_LCD_TOUCH_IO_SPI_STMPE610_CONFIG(EXAMPLE_PIN_NUM_TOUCH_CS);
-    // Attach the TOUCH to the SPI bus
-    ESP_ERROR_CHECK(esp_lcd_new_panel_io_spi((esp_lcd_spi_bus_handle_t)EXAMPLE_LCD_HOST, &tp_io_config, &tp_io_handle));
-
-    esp_lcd_touch_config_t tp_cfg = {
-        .x_max = EXAMPLE_LCD_H_RES,
-        .y_max = EXAMPLE_LCD_V_RES,
-        .rst_gpio_num = -1,
-        .int_gpio_num = -1,
-        .flags = {
-            .swap_xy = 0,
-            .mirror_x = 0,
-            .mirror_y = 0,
-        },
-    };
-    esp_lcd_touch_handle_t tp = NULL;
-
-#if CONFIG_EXAMPLE_LCD_TOUCH_CONTROLLER_STMPE610
-    ESP_LOGI(DISPLAY_TAG, "Initialize touch controller STMPE610");
-    ESP_ERROR_CHECK(esp_lcd_touch_new_spi_stmpe610(tp_io_handle, &tp_cfg, &tp));
-#endif // CONFIG_EXAMPLE_LCD_TOUCH_CONTROLLER_STMPE610
-
-    static lv_indev_t *indev;
-    indev = lv_indev_create(); // Input device driver (Touch)
-    lv_indev_set_type(indev, LV_INDEV_TYPE_POINTER);
-    lv_indev_set_display(indev, display);
-    lv_indev_set_user_data(indev, tp);
-    lv_indev_set_read_cb(indev, example_lvgl_touch_cb);
-#endif
-
-    ESP_LOGI(DISPLAY_TAG, "Create LVGL task");
-    xTaskCreate(example_lvgl_port_task, "LVGL", EXAMPLE_LVGL_TASK_STACK_SIZE, NULL, EXAMPLE_LVGL_TASK_PRIORITY, NULL);
-
-    ESP_LOGI(DISPLAY_TAG, "Display LVGL Meter Widget");
-    // Lock the mutex due to the LVGL APIs are not thread-safe
-    _lock_acquire(&lvgl_api_lock);
-    example_lvgl_demo_ui(display);
-    _lock_release(&lvgl_api_lock);
-
-    vTaskDelete(NULL);
+    return label1;
 }
