+++ src/app/display/display_frames.h
@@ -1,137 +1,10 @@
 #ifndef DISPLAY_FRAMES_H
 #define DISPLAY_FRAMES_H
+#include "lvgl.h"
 
-// 'SCAN logo', 128x128px
-// const unsigned char epd_bitmap_SCAN_logo[] = {
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xea, 0xaa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x55, 0x55, 0x57, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0x22, 0x22, 0x23, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0x55, 0x55, 0x55, 0x7f, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe8, 0xaf, 0xff, 0x8a, 0xbf, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x55, 0x7f, 0xff, 0xf5, 0x57, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x23, 0xff, 0xff, 0xfe, 0x2f, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x57, 0xff, 0xff, 0xff, 0x55, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf8, 0xbf, 0xff, 0xff, 0xff, 0xeb, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0x7f, 0xff, 0xff, 0xff, 0xd5, 0x7f, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe2, 0xff, 0xff, 0xff, 0xff, 0xfa, 0x7f, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd5, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x5f, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xab, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xbf, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x57, 0xff, 0xff, 0xff, 0xff, 0xff, 0x57, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x2f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xaf, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x5f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd7, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xbf, 0xff, 0xff, 0xff, 0xff, 0xff, 0xeb, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd5, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf3, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfa, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0x3f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xbf, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x3f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xaf, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0xff, 0x57, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfb, 0xff, 0xaf, 0xff, 0xff,
-//     0xff, 0xff, 0xdf, 0xff, 0xdf, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xff, 0xd7, 0xff, 0xff,
-//     0xff, 0xff, 0xef, 0xfe, 0xaf, 0xfe, 0xaf, 0xff, 0xfb, 0xff, 0xff, 0xfe, 0xff, 0xef, 0xff, 0xff,
-//     0xff, 0xfd, 0x55, 0xfd, 0x55, 0xfd, 0x57, 0xf5, 0xf5, 0xff, 0xf7, 0x7f, 0x7f, 0xd7, 0xff, 0xff,
-//     0xff, 0xff, 0xbb, 0xfb, 0xbb, 0xff, 0xbf, 0xfb, 0xfb, 0xff, 0xff, 0xbf, 0xbf, 0xe3, 0xff, 0xff,
-//     0xff, 0xfd, 0x55, 0xf5, 0x55, 0xfd, 0x57, 0xf5, 0xf5, 0xff, 0xff, 0x5f, 0x7f, 0xd7, 0xff, 0xff,
-//     0xff, 0xfa, 0xee, 0xfe, 0xee, 0xfe, 0xaf, 0xee, 0xff, 0xff, 0xff, 0xbf, 0xbf, 0xeb, 0xff, 0xff,
-//     0xff, 0xf5, 0x75, 0xf5, 0x75, 0x7d, 0x57, 0xf5, 0x75, 0xff, 0xf7, 0xd7, 0x5f, 0xf7, 0xff, 0xff,
-//     0xff, 0xfb, 0xfb, 0xfb, 0xfb, 0xff, 0xbf, 0xfb, 0xfb, 0xff, 0xfb, 0xef, 0xbf, 0xf3, 0xff, 0xff,
-//     0xff, 0xf5, 0x75, 0xf5, 0xf5, 0x7d, 0x57, 0xf5, 0x75, 0xff, 0xf5, 0xd7, 0xdf, 0xf5, 0xff, 0xff,
-//     0xff, 0xfa, 0xff, 0xee, 0xfe, 0xfe, 0xef, 0xea, 0xfb, 0xff, 0xfb, 0xef, 0xbf, 0xfb, 0xff, 0xff,
-//     0xff, 0xf5, 0x7f, 0xf5, 0x75, 0x7d, 0x57, 0xf5, 0x75, 0xff, 0xfd, 0xf7, 0xdf, 0xf5, 0xff, 0xff,
-//     0xff, 0xfb, 0xbf, 0xfb, 0xfb, 0xfb, 0xfb, 0xfb, 0xbb, 0xff, 0xfe, 0xfb, 0xbf, 0xf3, 0xff, 0xff,
-//     0xff, 0xfd, 0x5f, 0xf5, 0xff, 0xfd, 0x57, 0xf5, 0x55, 0xff, 0xfd, 0x77, 0xdf, 0xf5, 0xff, 0xff,
-//     0xff, 0xfe, 0xff, 0xea, 0xff, 0xfa, 0xef, 0xee, 0xbf, 0xff, 0xfe, 0xfb, 0xef, 0xfb, 0xff, 0xff,
-//     0xff, 0xfd, 0x57, 0xf5, 0xff, 0xf5, 0x57, 0xf5, 0x55, 0xff, 0xff, 0x77, 0xd7, 0xf5, 0xff, 0xff,
-//     0xff, 0xff, 0xbf, 0xfb, 0xff, 0xfb, 0xfb, 0xfb, 0xbb, 0xff, 0xfe, 0xfb, 0xef, 0xfb, 0xff, 0xff,
-//     0xff, 0xff, 0x55, 0xf5, 0xff, 0xf5, 0x55, 0xf5, 0x55, 0xff, 0xff, 0x75, 0xd7, 0xf5, 0xff, 0xff,
-//     0xff, 0xff, 0xab, 0xea, 0xff, 0xfe, 0xef, 0xeb, 0xeb, 0xff, 0xfe, 0xfb, 0xef, 0xfb, 0xff, 0xff,
-//     0xff, 0xff, 0x55, 0xf5, 0x7f, 0xf5, 0x75, 0xf5, 0x55, 0xff, 0xff, 0x77, 0xd7, 0xf5, 0xff, 0xff,
-//     0xff, 0xff, 0xfb, 0xfb, 0xff, 0xfb, 0xfb, 0xfb, 0xbb, 0xff, 0xfe, 0xfb, 0xef, 0xfb, 0xff, 0xff,
-//     0xff, 0xff, 0xd5, 0xf5, 0xf5, 0xf5, 0xd5, 0xf5, 0xd5, 0xff, 0xfd, 0x75, 0xdf, 0xf5, 0xff, 0xff,
-//     0xff, 0xff, 0xea, 0xee, 0xfa, 0xfe, 0xeb, 0xeb, 0xef, 0xff, 0xfe, 0xfb, 0xef, 0xfb, 0xff, 0xff,
-//     0xff, 0xf5, 0x75, 0x75, 0x75, 0x75, 0x55, 0xf5, 0xd5, 0xff, 0xfd, 0x77, 0xd7, 0xf5, 0xff, 0xff,
-//     0xff, 0xfb, 0xfb, 0xfb, 0xfb, 0xfb, 0xbb, 0xfb, 0xfb, 0xff, 0xfe, 0xf3, 0xbf, 0xfb, 0xff, 0xff,
-//     0xff, 0xf5, 0x75, 0x75, 0xf5, 0x75, 0x55, 0xf5, 0xd5, 0xff, 0xfd, 0xf7, 0xdf, 0xf5, 0xff, 0xff,
-//     0xff, 0xfe, 0xfa, 0xea, 0xfe, 0xea, 0xff, 0xeb, 0xeb, 0xff, 0xfb, 0xef, 0xbf, 0xfb, 0xff, 0xff,
-//     0xff, 0xf5, 0x75, 0x75, 0x75, 0x75, 0xf5, 0xf5, 0xf5, 0xff, 0xf5, 0xd7, 0x5f, 0xf5, 0xff, 0xff,
-//     0xff, 0xfb, 0xbb, 0xfb, 0xbb, 0xfb, 0xfb, 0xfb, 0xfb, 0xff, 0xfb, 0xef, 0xbf, 0xf3, 0xff, 0xff,
-//     0xff, 0xfd, 0x55, 0xf5, 0x55, 0xd5, 0xf5, 0xf5, 0xf5, 0xff, 0xff, 0xdf, 0x5f, 0xf5, 0xff, 0xff,
-//     0xff, 0xfe, 0xab, 0xfa, 0xab, 0xef, 0xfe, 0xeb, 0xff, 0xff, 0xff, 0xbf, 0xbf, 0xeb, 0xff, 0xff,
-//     0xff, 0xfd, 0x55, 0xfd, 0x57, 0xd5, 0xf5, 0x75, 0xf5, 0xff, 0xff, 0x5f, 0x7f, 0xf7, 0xff, 0xff,
-//     0xff, 0xff, 0xbf, 0xff, 0xbf, 0xfb, 0xfb, 0xfb, 0xfb, 0xff, 0xff, 0xbe, 0xff, 0xe3, 0xff, 0xff,
-//     0xff, 0xff, 0xdf, 0xff, 0xdf, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xd7, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xef, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0x7f, 0xd7, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfb, 0xff, 0xe7, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0xff, 0xd7, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x8f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x57, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xaf, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x9f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0x3f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfa, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xeb, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd5, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0x3f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe3, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd7, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xaf, 0xff, 0xff, 0xff, 0xff, 0xff, 0xaf, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x57, 0xff, 0xff, 0xff, 0xff, 0xf7, 0x5f, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xa3, 0xff, 0xff, 0xff, 0xff, 0xfe, 0x3f, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd5, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x7f, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xea, 0xff, 0xff, 0xff, 0xff, 0xfa, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0x7f, 0xff, 0xff, 0xff, 0xd5, 0x7f, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfa, 0x3f, 0xff, 0xff, 0xff, 0xa3, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0x57, 0xff, 0xff, 0xff, 0x55, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xab, 0xff, 0xff, 0xfe, 0xaf, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x55, 0x7f, 0xff, 0xf5, 0x5f, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe2, 0x3f, 0xff, 0xaa, 0x3f, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf5, 0x55, 0x55, 0x55, 0x7f, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xaa, 0xaa, 0x8b, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x55, 0x55, 0x57, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfa, 0x23, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-//     0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff};
+lv_obj_t *display_idle(lv_display_t *disp);
+lv_obj_t *display_check_in_failed(lv_display_t *disp);
+lv_obj_t *display_check_in_success(lv_display_t *disp);
+lv_obj_t *display_transmitting(lv_display_t *disp);
 
-void display_test(void *pvParameters);
-void display_test_1(void *pvParameters);
-#endif
+#endif // DISPLAY_FRAMES_H
