+++ src/app/main/CMakeLists.txt
@@ -1,6 +1,6 @@
 idf_component_register(
     SRCS "main.c"
     INCLUDE_DIRS "." "./include"
-    REQUIRES "display" "esp_driver_gpio" #"wifi" "pn532"  "keypad_driver" "NDEF" 
+    REQUIRES "display" "esp_driver_gpio" "pn532" "keypad_driver" "NDEF" # "wifi"
     PRIV_REQUIRES "nvs_flash"
     )
