+++ src/app/misc/CMakeLists.txt
@@ -1,0 +1,2 @@
+idf_component_register(SRCS "buzzer.c" "pir.c"
+                    INCLUDE_DIRS ".")
