+++ src/app/misc/buzzer.c
@@ -1,0 +1,1 @@
+#include "buzzer.h"
