+++ src/app/misc/buzzer.h
@@ -1,0 +1,4 @@
+#ifndef BUZZER_H
+#define BUZZER_H
+
+#endif // BUZZER_H
