+++ src/app/misc/pir.c
@@ -1,0 +1,1 @@
+#include "pir.h"
