+++ src/app/misc/pir.h
@@ -1,0 +1,4 @@
+#ifndef PIR_H
+#define PIR_H
+
+#endif // PIR_H
